Fixed centering to routePaths
