+++ src/components/sidebar/routeListView/RouteItem.tsx
@@ -1,9 +1,11 @@
 import { inject, observer } from 'mobx-react';
 import React from 'react';
+import L from 'leaflet';
 import classNames from 'classnames';
 import { FiInfo } from 'react-icons/fi';
 import Moment from 'moment';
 import { RouteListStore } from '~/stores/routeListStore';
+import { MapStore } from '~/stores/mapStore';
 import LineHelper from '~/util/lineHelper';
 import { dateToDateString } from '~/util/dateFormatHelper';
 import TransitTypeHelper from '~/util/TransitTypeHelper';
@@ -19,21 +21,43 @@ import * as s from './routeItem.scss';
 
 interface IRouteItemProps {
     routeListStore?: RouteListStore;
+    mapStore?: MapStore;
     route: IRoute;
 }
 
-@inject('routeListStore')
+@inject('routeListStore', 'mapStore')
 @observer
 class RouteItem extends React.Component<IRouteItemProps> {
     async componentDidMount() {
-        this.props.route.routePaths.forEach((routePath, index) => {
-            // Make two first route paths visible by default
+        let index = 0;
+        for (const routePath of this.props.route.routePaths) {
             if (index < 2) {
-                this.props.routeListStore!.toggleRoutePathVisibility(
+                await this.props.routeListStore!.toggleRoutePathVisibility(
                     routePath.internalId
                 );
             }
+            index += 1;
+        }
+        this.calculateBounds();
+    }
+
+    private calculateBounds() {
+        const routes = this.props.routeListStore!.routes;
+        const bounds: L.LatLngBounds = new L.LatLngBounds([]);
+
+        routes.forEach(route => {
+            route.routePaths.forEach(routePath => {
+                routePath.routePathLinks.forEach(routePathLink => {
+                    routePathLink.geometry.forEach(pos => {
+                        bounds.extend(pos);
+                    });
+                });
+            });
         });
+
+        if (bounds.isValid()) {
+            this.props.mapStore!.setMapBounds(bounds);
+        }
     }
 
     private closeRoute = () => {
@@ -104,12 +128,14 @@ class RouteItem extends React.Component<IRouteItemProps> {
                 this.props.routeListStore!.toggleRoutePathVisibility(
                     routePath.internalId
                 );
+                this.calculateBounds();
             };
 
             const openRoutePathView = () => {
                 const routePathViewLink = routeBuilder
                     .to(subSites.routePath)
-                    .toTarget(':id',
+                    .toTarget(
+                        ':id',
                         [
                             routePath.routeId,
                             Moment(routePath.startTime).format(
